% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{addNode}
\alias{addNode}
\title{addNode}
\usage{
addNode(circleAFMImage, nodeDT, filterIndex)
}
\arguments{
\item{circleAFMImage}{a \code{\link{AFMImage}}}

\item{nodeDT}{nodeDT a data.table lon lat circleRadius}

\item{filterIndex}{an integer}
}
\value{
an \code{\link{AFMImage}}
}
\description{
add a node to an AFMImage
}
\author{
M.Beauvais
}
