% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImage.R
\name{filterAFMImage}
\alias{filterAFMImage}
\title{filter the heights of an AFMImage with a minimun and a maximum value}
\usage{
filterAFMImage(AFMImage, Min, Max)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{Min}{the minimun height value to keep}

\item{Max}{the maximun height value to keep}
}
\value{
an \code{\link{AFMImage}}
}
\description{
\code{filterAFMImage} returns a filtered AFMImage
}
\author{
M.Beauvais
}
