% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImage.R
\name{isBinary}
\alias{isBinary}
\title{has the AFM Image heights of 0 or 1}
\usage{
isBinary(AFMImage)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}
}
\value{
a boolean
}
\description{
\code{isBinary} returns TRUE is the heights of the AFMImage is 0 or 1
}
\author{
M.Beauvais
}
