% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImage.R
\name{makeBinaryAFMImage}
\alias{makeBinaryAFMImage}
\title{make a binary AFMImage setting all the heights different to 0 to 1.}
\usage{
makeBinaryAFMImage(AFMImage)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}
}
\value{
an \code{\link{AFMImage}}
}
\description{
\code{makeBinaryAFMImage} returns a binary AFMImage
}
\author{
M.Beauvais
}
