% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImage.R
\name{multiplyHeightsAFMImage}
\alias{multiplyHeightsAFMImage}
\title{multiply the heights of an AFMImage}
\usage{
multiplyHeightsAFMImage(AFMImage, multiplier)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{multiplier}{the number to multiply the heights with}
}
\value{
an \code{\link{AFMImage}}
}
\description{
\code{multiplyHeightsAFMImage} returns a simplified AFMImage
}
\examples{
\dontrun{
data(AFMImageOfAluminiumInterface)
newAFMImage<-multiplyHeightsAFMImage(AFMImageOfAluminiumInterface,10)
displayIn3D(newAFMImage,noLight=TRUE)
}

}
\author{
M.Beauvais
}
