% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM3DPrinter.R
\docType{class}
\name{AFMImage3DModelAnalysis-class}
\alias{AFMImage3DModelAnalysis-class}
\alias{AFMImage3DModelAnalysis}
\title{AFM image Power Spectrum Density analysis class}
\description{
\code{AFMImage3DModelAnalysis}
}
\section{Slots}{

\describe{
\item{\code{f1}}{a face of the 3D model}

\item{\code{f2}}{a face of the 3D model}

\item{\code{f3}}{a face of the 3D model}

\item{\code{f4}}{a face of the 3D model}
}}

\author{
M.Beauvais
}
