% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImageAnalyser.R
\name{initialize,AFMImageAnalyser-method}
\alias{initialize,AFMImageAnalyser-method}
\title{Constructor method of AFMImageAnalyser Class.}
\usage{
\S4method{initialize}{AFMImageAnalyser}(
  .Object,
  AFMImage,
  variogramAnalysis,
  psdAnalysis,
  fdAnalysis,
  gaussianMixAnalysis,
  networksAnalysis,
  threeDimensionAnalysis,
  mean,
  variance,
  TotalRrms,
  Ra,
  fullfilename
)
}
\arguments{
\item{.Object}{an AFMImageAnalyser object}

\item{AFMImage}{an \code{AFMImage}}

\item{variogramAnalysis}{\code{\link{AFMImageVariogramAnalysis}}}

\item{psdAnalysis}{\code{\link{AFMImagePSDAnalysis}}}

\item{fdAnalysis}{\code{\link{AFMImageFractalDimensionsAnalysis}}}

\item{gaussianMixAnalysis}{\code{\link{AFMImageGaussianMixAnalysis}}}

\item{networksAnalysis}{\code{\link{AFMImageNetworksAnalysis}}}

\item{threeDimensionAnalysis}{\code{\link{AFMImage3DModelAnalysis}}}

\item{mean}{the mean of heights of the \code{\link{AFMImage}}}

\item{variance}{the variance of heights of the \code{\link{AFMImage}}}

\item{TotalRrms}{the total Root Mean Square Roughness of the \code{\link{AFMImage}} calculated from variance}

\item{Ra}{mean roughness or mean of absolute values of heights}

\item{fullfilename}{to be removed?}
}
\description{
Constructor method of AFMImageAnalyser Class.
}
