% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImageAnalyser.R
\name{AFMImageAnalyser-class}
\alias{AFMImageAnalyser-class}
\alias{AFMImageAnalyser}
\title{AFM image analyser class}
\usage{
AFMImageAnalyser(AFMImage)

AFMImageAnalyser(AFMImage)
}
\arguments{
\item{AFMImage}{an \code{AFMImage}}
}
\description{
A S4 class to handle the analysis of one AFM Image.
}
\section{Slots}{

\describe{
\item{\code{AFMImage}}{\code{\link{AFMImage}} to be analysed}

\item{\code{variogramAnalysis}}{\code{\link{AFMImageVariogramAnalysis}}}

\item{\code{psdAnalysis}}{\code{\link{AFMImagePSDAnalysis}}}

\item{\code{fdAnalysis}}{\code{\link{AFMImageFractalDimensionsAnalysis}}}

\item{\code{gaussianMixAnalysis}}{\code{\link{AFMImageGaussianMixAnalysis}}}

\item{\code{networksAnalysis}}{\code{\link{AFMImageNetworksAnalysis}}}

\item{\code{mean}}{the mean of heights of the \code{\link{AFMImage}}}

\item{\code{variance}}{the variance of heights of the \code{\link{AFMImage}}}

\item{\code{TotalRrms}}{the total Root Mean Square Roughness of the \code{\link{AFMImage}} calculated from variance}

\item{\code{Ra}}{mean roughness or mean of absolute values of heights}

\item{\code{fullfilename}}{to be removed ?}

\item{\code{updateProgress}}{a function to update a graphical user interface}
}}

\author{
M.Beauvais
}
