% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImage.R
\name{AFMImageCollagenNetwork}
\alias{AFMImageCollagenNetwork}
\title{AFM image sample}
\description{
A real dataset containing an \code{\link{AFMImage}} of a collagen network.
The image is made of 192*192 samples of a 1500 nm * 1500 nm surface.
samplesperline=192
lines=192
hscansize=1500
vscansize=1500
}
