% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMFractalDimensionAnalyser.R
\name{AFMImageFractalDimensionsAnalysis-class}
\alias{AFMImageFractalDimensionsAnalysis-class}
\alias{AFMImageFractalDimensionsAnalysis}
\alias{initialize,AFMImageFractalDimensionsAnalysis-method}
\alias{fractalDimensionMethods,AFMImageFractalDimensionsAnalysis-method}
\alias{fractalDimensionMethods}
\title{AFM image fractal dimensions analysis class}
\usage{
AFMImageFractalDimensionsAnalysis()

\S4method{initialize}{AFMImageFractalDimensionsAnalysis}(.Object, fractalDimensionMethods, csvFullfilename)

AFMImageFractalDimensionsAnalysis()

fractalDimensionMethods(object)

\S4method{fractalDimensionMethods}{AFMImageFractalDimensionsAnalysis}(object)
}
\arguments{
\item{.Object}{an AFMImageFractalDimensionsAnalysis Class}

\item{fractalDimensionMethods}{a list of \code{\link{AFMImageFractalDimensionMethod}}}

\item{csvFullfilename}{To be removed ?}

\item{object}{a \code{\link{AFMImageFractalDimensionsAnalysis}}}
}
\description{
A S4 class to handle the fractal dimension calculation with several fractal dimension methods
}
\section{Slots}{

\describe{
\item{\code{fractalDimensionMethods}}{a list of \code{\link{AFMImageFractalDimensionMethod}}}

\item{\code{csvFullfilename}}{To be removed ?}

\item{\code{updateProgress}}{a function to update a graphical user interface}
}}

\author{
M.Beauvais
}
