% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMGaussianMixAnalyser.R
\name{AFMImageGaussianMixAnalysis-class}
\alias{AFMImageGaussianMixAnalysis-class}
\alias{AFMImageGaussianMixAnalysis}
\alias{initialize,AFMImageGaussianMixAnalysis-method}
\alias{summaryMixture,AFMImageGaussianMixAnalysis-method}
\alias{summaryMixture}
\alias{eachComponentsCounts,AFMImageGaussianMixAnalysis-method}
\alias{eachComponentsCounts}
\alias{densityCurvesAllHeights,AFMImageGaussianMixAnalysis-method}
\alias{densityCurvesAllHeights}
\alias{tcdfsEcdfsCheck,AFMImageGaussianMixAnalysis-method}
\alias{tcdfsEcdfsCheck}
\alias{gaussianMix,AFMImageGaussianMixAnalysis-method}
\alias{gaussianMix}
\alias{minGaussianMix,AFMImageGaussianMixAnalysis-method}
\alias{minGaussianMix}
\alias{maxGaussianMix,AFMImageGaussianMixAnalysis-method}
\alias{maxGaussianMix}
\alias{epsilonGaussianMix,AFMImageGaussianMixAnalysis-method}
\alias{epsilonGaussianMix}
\title{AFM image Gaussian Mix analysis class}
\usage{
AFMImageGaussianMixAnalysis()

\S4method{initialize}{AFMImageGaussianMixAnalysis}(.Object)

AFMImageGaussianMixAnalysis()

summaryMixture(object)

\S4method{summaryMixture}{AFMImageGaussianMixAnalysis}(object)

eachComponentsCounts(object)

\S4method{eachComponentsCounts}{AFMImageGaussianMixAnalysis}(object)

tcdfsEcdfsCheck(object)

densityCurvesAllHeights(object)

\S4method{densityCurvesAllHeights}{AFMImageGaussianMixAnalysis}(object)

tcdfsEcdfsCheck(object)

\S4method{tcdfsEcdfsCheck}{AFMImageGaussianMixAnalysis}(object)

gaussianMix(object)

\S4method{gaussianMix}{AFMImageGaussianMixAnalysis}(object)

minGaussianMix(object)

\S4method{minGaussianMix}{AFMImageGaussianMixAnalysis}(object)

maxGaussianMix(object)

\S4method{maxGaussianMix}{AFMImageGaussianMixAnalysis}(object)

epsilonGaussianMix(object)

\S4method{epsilonGaussianMix}{AFMImageGaussianMixAnalysis}(object)
}
\arguments{
\item{.Object}{an AFMImageGaussianMixAnalysis object}

\item{object}{a \code{\link{AFMImageGaussianMixAnalysis}}}
}
\description{
\code{AFMImageGaussianMixAnalysis} handles an \code{\link{AFMImage}} Gaussian mix of heights analysis
}
\section{Slots}{

\describe{
\item{\code{minGaussianMix}}{the minimum number of components to calculate}

\item{\code{maxGaussianMix}}{the maximum number of components to calculate}

\item{\code{epsilonGaussianMix}}{the convergence criterion}

\item{\code{gaussianMix}}{a data.table to store the calculated Gaussian mixes}

\item{\code{summaryMixture}}{a data.table to summaryse the mixtures}

\item{\code{tcdfsEcdfsCheck}}{an array to store the points to draw tcdfs ecdfs check}

\item{\code{densityCurvesAllHeights}}{an array to store the points to draw the density curves}

\item{\code{eachComponentsCounts}}{an array to store the points to draw counts of each components}

\item{\code{updateProgress}}{a function to update a graphical user interface}
}}

\author{
M.Beauvais
}
