% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImage.R
\name{AFMImageOfAluminiumInterface}
\alias{AFMImageOfAluminiumInterface}
\title{AFM image sample}
\description{
A real dataset containing an \code{\link{AFMImage}} of an Aluminium interface.
The image is made of 512*512 samples of a 1000 nm * 1000 nm surface.
samplesperline=512
lines=512
hscansize=1000
vscansize=1000
}
\author{
J.Landoulsi, I.Liascukiene
}
