% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImage.R
\name{AFMImageOfOnePeak}
\alias{AFMImageOfOnePeak}
\title{AFM image sample}
\description{
A fake dataset containing a manually generated \code{\link{AFMImage}} (one peak positioned on the surface).
The image is made of 128*128 samples of a 128 nm * 128 nm surface.
samplesperline= 128
lines= 128
hscansize= 128
vscansize= 128
}
