% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMPSDAnalyser.R
\name{AFMImagePSDAnalysis-class}
\alias{AFMImagePSDAnalysis-class}
\alias{AFMImagePSDAnalysis}
\alias{initialize,AFMImagePSDAnalysis-method}
\alias{psd1d_breaks,AFMImagePSDAnalysis-method}
\alias{psd1d_breaks}
\alias{psd2d_maxHighLengthScale,AFMImagePSDAnalysis-method}
\alias{psd2d_maxHighLengthScale}
\alias{psd2d_truncHighLengthScale,AFMImagePSDAnalysis-method}
\alias{psd2d_truncHighLengthScale}
\alias{psd1d,AFMImagePSDAnalysis-method}
\alias{psd1d}
\alias{psd2d,AFMImagePSDAnalysis-method}
\alias{psd2d}
\alias{roughnessAgainstLengthscale,AFMImagePSDAnalysis-method}
\alias{roughnessAgainstLengthscale}
\alias{intersections,AFMImagePSDAnalysis-method}
\alias{intersections}
\title{AFM image Power Spectrum Density analysis class}
\usage{
AFMImagePSDAnalysis()

\S4method{initialize}{AFMImagePSDAnalysis}(.Object)

AFMImagePSDAnalysis()

psd1d_breaks(object)

\S4method{psd1d_breaks}{AFMImagePSDAnalysis}(object)

psd2d_maxHighLengthScale(object)

\S4method{psd2d_maxHighLengthScale}{AFMImagePSDAnalysis}(object)

psd2d_truncHighLengthScale(object)

\S4method{psd2d_truncHighLengthScale}{AFMImagePSDAnalysis}(object)

psd1d(object)

\S4method{psd1d}{AFMImagePSDAnalysis}(object)

psd2d(object)

\S4method{psd2d}{AFMImagePSDAnalysis}(object)

roughnessAgainstLengthscale(object)

\S4method{roughnessAgainstLengthscale}{AFMImagePSDAnalysis}(object)

intersections(object)

\S4method{intersections}{AFMImagePSDAnalysis}(object)
}
\arguments{
\item{.Object}{an AFMImagePSDAnalysis object}

\item{object}{a \code{\link{AFMImagePSDAnalysis}}}
}
\description{
\code{AFMImagePSDAnalysis} handles an \code{\link{AFMImage}} roughness against lenghscale analysis
}
\section{Slots}{

\describe{
\item{\code{roughnessAgainstLengthscale}}{a data.table to store the roughness against lengthscale data}

\item{\code{intersections}}{a list to store the lengthscales values as the intersections between slopes and the sill in roughness against lenghscale graph}

\item{\code{updateProgress}}{a function to update a graphical user interface}
}}

\author{
M.Beauvais
}
