% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMVariogramAnalyser.R
\docType{methods}
\name{getDTModelEvaluation}
\alias{getDTModelEvaluation}
\alias{getDTModelEvaluation,AFMImageVariogramAnalysis-method}
\title{getDTModelEvaluation method}
\usage{
getDTModelEvaluation(AFMImageVariogramAnalysis)

\S4method{getDTModelEvaluation}{AFMImageVariogramAnalysis}(AFMImageVariogramAnalysis)
}
\arguments{
\item{AFMImageVariogramAnalysis}{an AFMImageVariogramAnalysis object}
}
\description{
getDTModelEvaluation method
}
