% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMVariogramAnalyser.R
\docType{methods}
\name{getDTModelSillRange}
\alias{getDTModelSillRange}
\alias{getDTModelSillRange,AFMImageVariogramAnalysis-method}
\title{getDTModelSillRange method}
\usage{
getDTModelSillRange(AFMImageVariogramAnalysis)

\S4method{getDTModelSillRange}{AFMImageVariogramAnalysis}(AFMImageVariogramAnalysis)
}
\arguments{
\item{AFMImageVariogramAnalysis}{an AFMImageVariogramAnalysis object}
}
\description{
getDTModelSillRange method
}
