% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMVariogramAnalyser.R
\docType{methods}
\name{updateProgress}
\alias{updateProgress}
\alias{updateProgress,AFMImageVariogramAnalysis-method}
\title{updateProgress}
\usage{
updateProgress(AFMImageVariogramAnalysis, value, detail, message)
}
\arguments{
\item{AFMImageVariogramAnalysis}{an \code{\link{AFMImageVariogramAnalysis}}}

\item{value}{shiny progress bar value}

\item{detail}{shiny progress bar detail}

\item{message}{shiny progress bar message}
}
\description{
is a function used by a GUI such as shiny GUI
}
