% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMVariogramAnalyser.R
\name{AFMImageVariogramModel-class}
\alias{AFMImageVariogramModel-class}
\alias{AFMImageVariogramModel}
\alias{initialize,AFMImageVariogramModel-method}
\title{AFM Image Variogram Model class}
\usage{
AFMImageVariogramModel()

\S4method{initialize}{AFMImageVariogramModel}(
  .Object,
  model,
  fit.v = data.table(),
  mykrige,
  res = data.table(),
  cor,
  press,
  sill,
  imageFullfilename
)

AFMImageVariogramModel()
}
\arguments{
\item{.Object}{an AFMImageVariogramModel object}

\item{model}{the variogram model name}

\item{fit.v}{the values from the \code{\link[gstat]{fit.variogram}} function in the gstat package}

\item{mykrige}{the values from the \code{\link[gstat]{krige}} function in the gstat library}

\item{res}{a data.table to store: (cor) the correlation between the predicted sample and the real sample (press) the sum of the square of the differences between real and predicted values for each point of the sample}

\item{cor}{to be removed ?}

\item{press}{to be removed ?}

\item{sill}{to be removed ?}

\item{imageFullfilename}{to be removed ?}
}
\description{
\code{AFMImageVariogramModel}stores the evaluation of one experimental variogram model
}
\section{Slots}{

\describe{
\item{\code{model}}{the variogram model name}

\item{\code{fit.v}}{the values from the \code{\link[gstat]{fit.variogram}} function in the gstat package}

\item{\code{mykrige}}{the values from the \code{\link[gstat]{krige}} function in the gstat library}

\item{\code{res}}{a data.table to store: (cor) the correlation between the predicted sample and the real sample (press) the sum of the square of the differences between real and predicted values for each point of the sample}

\item{\code{cor}}{to be removed ?}

\item{\code{press}}{to be removed ?}

\item{\code{sill}}{to be removed ?}

\item{\code{imageFullfilename}}{to be removed ?}
}}

\author{
M.Beauvais
}
