% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMVariogramAnalyser.R
\docType{methods}
\name{AFMImageVariogramSlopesAnalysis-class}
\alias{AFMImageVariogramSlopesAnalysis}
\alias{AFMImageVariogramSlopesAnalysis-class}
\alias{initialize,AFMImageVariogramSlopesAnalysis-method}
\title{AFM Image psd slope analysis}
\usage{
AFMImageVariogramSlopesAnalysis()

\S4method{initialize}{AFMImageVariogramSlopesAnalysis}(.Object)

AFMImageVariogramSlopesAnalysis()
}
\arguments{
\item{.Object}{an AFMImageVariogramSlopesAnalysis object}
}
\description{
\code{AFMImageVariogramSlopesAnalysis} stores the analysis of the second slope in roughness against lenghtscale

Constructor method of AFMImageVariogramSlopesAnalysis Class.

Wrapper function AFMImageVariogramSlopesAnalysis
}
\section{Slots}{

\describe{
\item{\code{intersection_sill}}{to be removed ?}

\item{\code{sill}}{to be removed ?}

\item{\code{slope}}{to be removed ?}

\item{\code{yintersept}}{to be removed ?}
}}
\author{
M.Beauvais
}

