% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{AreNodesConnected}
\alias{AreNodesConnected}
\title{check if nodes represented by circles are connected. The function defines all the possible segments between the circles and check if at least one segment exists.}
\usage{
AreNodesConnected(binaryAFMImage, center1, radius1, center2, radius2)
}
\arguments{
\item{binaryAFMImage}{a binary \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{center1}{the center of the circle with center$lon as the x coordinates and center$lat as the y coordinates}

\item{radius1}{the radius of the circle}

\item{center2}{the center of the circle with center$lon as the x coordinates and center$lat as the y coordinates}

\item{radius2}{the radius of the circle}
}
\value{
TRUE if the nodes are connected
}
\description{
check if nodes represented by circles are connected. The function defines all the possible segments between the circles and check if at least one segment exists.
}
\author{
M.Beauvais
}
