% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM3DPrinter.R
\name{calculate3DModel}
\alias{calculate3DModel}
\alias{calculate3DModel,AFMImage3DModelAnalysis-method}
\alias{calculate3DModel,AFMImage-method}
\title{Calculate the 3D model for 3D printing}
\usage{
calculate3DModel(AFMImage3DModelAnalysis, AFMImage)

\S4method{calculate3DModel}{AFMImage3DModelAnalysis}(AFMImage3DModelAnalysis, AFMImage)
}
\arguments{
\item{AFMImage3DModelAnalysis}{n \code{\link{AFMImage3DModelAnalysis}} to store the setup and results of PSD analysis}

\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}
}
\description{
\code{calculate3DModel} update  \code{\link{AFMImage3DModelAnalysis}}
}
\author{
M.Beauvais
}
