% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculateGaussianMixture}
\alias{calculateGaussianMixture}
\title{Calculate Gaussian Mixture with two components from the AFM Image.}
\usage{
calculateGaussianMixture(AFMImage)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}
}
\description{
\code{calculateGaussianMixture} return a data.table containing the result of the Gaussian Mixture and result of the test
}
\examples{
\dontrun{
library(AFM)
data(AFMImageOfNetworks)
mixtureCharacteristics<-calculateGaussianMixture(AFMImageOfNetworks)
print(mixtureCharacteristics)
}
}
\author{
M.Beauvais
}
