% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculateHolesCharacteristics}
\alias{calculateHolesCharacteristics}
\title{get the networks parameters}
\usage{
calculateHolesCharacteristics(AFMImageNetworksAnalysis)
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}
}
\value{
a data.table with all the parameters
}
\description{
Calculate the holes characteristics
}
\author{
M.Beauvais
}
