% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculateIgraph}
\alias{calculateIgraph}
\title{Calculate iGraph from AFMImage}
\usage{
calculateIgraph(AFMImage, AFMImageNetworksAnalysis)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{AFMImageNetworksAnalysis}{an \code{\link{AFMImageNetworksAnalysis}} from Atomic Force Microscopy}
}
\description{
\code{calculateIgraph} return
}
\author{
M.Beauvais
}
