% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculateNetworkParameters}
\alias{calculateNetworkParameters}
\title{get the networks parameters}
\usage{
calculateNetworkParameters(AFMImageNetworksAnalysis, AFMImage)
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{AFMImage}{a \code{\link{AFMImage}}}
}
\value{
a data.table with all the parameters
}
\description{
Calculate and return the networks parameters
}
\author{
M.Beauvais
}
