% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculateNetworks}
\alias{calculateNetworks}
\alias{calculateNetworks,AFMImageNetworksAnalysis-method}
\alias{calculateNetworks,AFMImage-method}
\title{Calculate networks on the surface}
\usage{
calculateNetworks(AFMImageNetworksAnalysis, AFMImage)

\S4method{calculateNetworks}{AFMImageNetworksAnalysis}(AFMImageNetworksAnalysis, AFMImage)
}
\arguments{
\item{AFMImageNetworksAnalysis}{n \code{\link{AFMImageNetworksAnalysis}} to store the results of networks analysis}

\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}
}
\description{
\code{calculateNetworks} update  \code{\link{AFMImageNetworksAnalysis}}
}
\author{
M.Beauvais
}
