% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculateNetworkSkeleton}
\alias{calculateNetworkSkeleton}
\title{calculateNetworkSkeleton}
\usage{
calculateNetworkSkeleton(AFMImage, AFMImageNetworksAnalysis)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{AFMImageNetworksAnalysis}{an \code{\link{AFMImageNetworksAnalysis}} from Atomic Force Microscopy}
}
\description{
\code{calculateNetworkSkeleton} return
}
\author{
M.Beauvais
}
