% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculatePhysicalDistanceFromPath}
\alias{calculatePhysicalDistanceFromPath}
\title{calculate the physical distances between nodes}
\usage{
calculatePhysicalDistanceFromPath(AFMImage, pathVidVector)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{pathVidVector}{a network path}
}
\value{
the physical distance the extrmities of the path
}
\description{
calculate the physical distances between nodes
}
\author{
M.Beauvais
}

