% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculateShortestPaths}
\alias{calculateShortestPaths}
\title{calculate the  shortest path between adjacent nodes}
\usage{
calculateShortestPaths(..., AFMImageNetworksAnalysis)
}
\arguments{
\item{...}{cl: a cluster object from the parallel package}

\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}
}
\description{
Calculate the shortest path between all nodes of degree different to 2
that are connected with nodes of degree equal to 2
Calculate the distance between the above nodes.
}
\author{
M.Beauvais
}
