% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{canBeRemoved}
\alias{canBeRemoved}
\title{canBeRemoved}
\usage{
canBeRemoved(vertexId, g, allVertices, DEGREE_LIMIT_FOR_CANDIDATE_VERTICE)
}
\arguments{
\item{vertexId}{a vertex id}

\item{g}{a igraph}

\item{allVertices}{list of all vertices}

\item{DEGREE_LIMIT_FOR_CANDIDATE_VERTICE}{degree}
}
\description{
\code{canBeRemoved} return
}
\author{
M.Beauvais
}
