% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{createGraph}
\alias{createGraph}
\title{create the igraph weighted graph from the nodes and edges}
\usage{
createGraph(AFMImageNetworksAnalysis)
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}
}
\description{
create the igraph weighted graph from the nodes and edges
}
\author{
M.Beauvais
}
