% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{displayColoredNetworkWithVerticesSize}
\alias{displayColoredNetworkWithVerticesSize}
\title{displayColoredNetworkWithVerticesSize}
\usage{
displayColoredNetworkWithVerticesSize(AFMImageNetworksAnalysis, fullfilename)
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{fullfilename}{a directory plus filename for export}
}
\description{
display network
}
\author{
M.Beauvais
}
