% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{displaygridIgraphPlot}
\alias{displaygridIgraphPlot}
\title{display the network of nodes and edges}
\usage{
displaygridIgraphPlot(AFMImageNetworksAnalysis)
}
\arguments{
\item{AFMImageNetworksAnalysis}{an \code{\link{AFMImageNetworksAnalysis}}}
}
\description{
display the network of nodes and edges
}
\author{
M.Beauvais
}
