% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{displaygridIgraphPlotFromEdges}
\alias{displaygridIgraphPlotFromEdges}
\title{display the network of nodes and edges}
\usage{
displaygridIgraphPlotFromEdges(AFMImage, edges, isolates)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{edges}{list of edges}

\item{isolates}{list of isolated edges}
}
\description{
display the network of nodes and edges
}
\author{
M.Beauvais
}
