% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM3DPrinter.R
\name{displayIn3D}
\alias{displayIn3D}
\title{Display a 3D image of an AFMImage and store it on disk.}
\usage{
displayIn3D(AFMImage, width, fullfilename, changeViewpoint, noLight)
}
\arguments{
\item{AFMImage}{the AFM image to be displayed in three dimensions.}

\item{width}{(optional)  width of the image. Default is 512 pixels. Note: width can't be superior to screen resolution.}

\item{fullfilename}{(optional) the directory and filename to save the png of the 3D image. If this variable is missing, the function will not save on disk the 3D image.}

\item{changeViewpoint}{(optional) if TRUE, the viewpoint is changed. Default is TRUE.}

\item{noLight}{if TRUE, the ligth is set off}
}
\description{
Display a 3D image of an AFMImage and store it on disk if fullfilename variable is set.
It uses the \code{\link{rgl}} package.
}
\author{
M.Beauvais
}
