% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMGaussianMixAnalyser.R
\name{dnormalmix}
\alias{dnormalmix}
\title{dnormalmix density of a mixture of normals}
\usage{
dnormalmix(x, mixture, log = FALSE)
}
\arguments{
\item{x}{a vector of quantiles}

\item{mixture}{a gaussian mixture}

\item{log}{perform a log transsformation of the result}
}
\description{
dnormalmix density of a mixture of normals
}
