% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{existsEdge}
\alias{existsEdge}
\title{Does an edge exist ?}
\usage{
existsEdge(AFMImage, vertexId)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{vertexId}{the vertex id}
}
\description{
\code{existsEdge} return TRUE if an edge exists for this vertex id
}
\author{
M.Beauvais
}
