% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{existsSegment}
\alias{existsSegment}
\title{existsSegment checks if a segment exists in an AFMImage; check if all the heights at the segment coordinates are different to zero.}
\usage{
existsSegment(AFMImage, segment)
}
\arguments{
\item{AFMImage}{a \code{\link{AFMImage}} from Atomic Force Microscopy or a binary \code{\link{AFMImage}}}

\item{segment}{a data.table coming from the getBresenham2Dsegment - x and y should start from 1,1 #TODO Segment class}
}
\value{
TRUE if all the heights of the segment are different from zero
}
\description{
\code{existsSegment} return a boolean
}
\author{
M.Beauvais
}
