% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMReportMaker.R
\name{generateAFMImageReport}
\alias{generateAFMImageReport}
\title{Generate an analysis report from an AFMImageAnalyser object}
\usage{
generateAFMImageReport(AFMImageAnalyser, reportFullfilename, isCheckReport)
}
\arguments{
\item{AFMImageAnalyser}{an \code{\link{AFMImageAnalyser}} to be used to produce report}

\item{reportFullfilename}{location on disk where to save the generated report}

\item{isCheckReport}{TRUE to generate a check report must be generated, FALSE to generate a full report}
}
\description{
\code{generateAFMImageReport} generates a report from an AFMImageAnalyser object
}
\author{
M.Beauvais
}
