% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{generatePolygonEnvelope}
\alias{generatePolygonEnvelope}
\title{generatePolygonEnvelope}
\usage{
generatePolygonEnvelope(AFMImageNetworksAnalysis, centers, radius)
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{centers}{a matrix ?}

\item{radius}{a vector of radius}
}
\value{
a polygon
}
\description{
generate a convex polygon from circles
}
\author{
M.Beauvais
}
