% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM3DPrinter.R
\name{get3DImageFullfilename}
\alias{get3DImageFullfilename}
\title{get 3D image full filename}
\usage{
get3DImageFullfilename(exportDirectory, imageName)
}
\arguments{
\item{exportDirectory}{a diretcory to export image}

\item{imageName}{the image name}
}
\description{
get 3D image full filename
}
\author{
M.Beauvais
}
