% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getAllPointsToRemove}
\alias{getAllPointsToRemove}
\title{getAllPointsToRemove}
\usage{
getAllPointsToRemove(AFMImageNetworksAnalysis, envelope)
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{envelope}{an envelope of points ?}
}
\value{
a data.table of points
}
\description{
get the points inside envelope
}
\author{
M.Beauvais
}
