% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getAngle}
\alias{getAngle}
\title{calculate the angle between two vectors}
\usage{
getAngle(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the angle between the vectors
}
\description{
calculate the angle between two vectors
}
\author{
M.Beauvais
}
