% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMVariogramAnalyser.R
\name{getAutoIntersectionForOmnidirectionalVariogram}
\alias{getAutoIntersectionForOmnidirectionalVariogram}
\title{Calculate slopes and intersections in variogram
\code{getAutoIntersectionForOmnidirectionalVariogram} returns the slope in the omnidirectional variograms}
\usage{
getAutoIntersectionForOmnidirectionalVariogram(AFMImageAnalyser)
}
\arguments{
\item{AFMImageAnalyser}{an \code{\link{AFMImageAnalyser}}}
}
\value{
an \code{\link{omniVariogramSlopeAnalysis}}
}
\description{
Calculate slopes and intersections in variogram
\code{getAutoIntersectionForOmnidirectionalVariogram} returns the slope in the omnidirectional variograms
}
\author{
M.Beauvais
}
