% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getBresenham2DSegment}
\alias{getBresenham2DSegment}
\title{get a segment of points thanks to Bresenham line algorithm}
\usage{
getBresenham2DSegment(x1, y1, x2, y2)
}
\arguments{
\item{x1}{abscissa coordinates of the first point}

\item{y1}{ordinate coordinates of the first point}

\item{x2}{abscissa coordinates of the second point}

\item{y2}{ordinate coordinates of the second point}
}
\value{
a data.table of points - data.table(x, y)
}
\description{
\code{getBresenham2DSegment} return the Bresenham segment in 2D from extremities coordinates
}
\author{
M.Beauvais
}
