% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getCircleSpatialPoints}
\alias{getCircleSpatialPoints}
\title{get the spatial points on the circle including the center of the circle}
\usage{
getCircleSpatialPoints(binaryAFMImage, center, circleRadius)
}
\arguments{
\item{binaryAFMImage}{a binary \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{center}{the center of the circle with center$lon as the x coordinates and center$lat as the y coordinates}

\item{circleRadius}{the radius of the circle}
}
\value{
a \code{\link{SpatialPoints}} object of all the points of the circle including the center of the circle
}
\description{
get the spatial points on the circle including the center of the circle
}
\author{
M.Beauvais
}
