% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMPSDAnalyser.R
\name{getIntersectionForRoughnessAgainstLengthscale}
\alias{getIntersectionForRoughnessAgainstLengthscale}
\title{get the intersection between tangente and plateau}
\usage{
getIntersectionForRoughnessAgainstLengthscale(
  AFMImageAnalyser,
  minValue,
  maxValue,
  second_slope = FALSE
)
}
\arguments{
\item{AFMImageAnalyser}{an \code{\link{AFMImageAnalyser}} to get Roughness against lenghtscale calculation}

\item{minValue}{index of the lowest point to be used for the tangent}

\item{maxValue}{index of the highest point to be used for the tangent}

\item{second_slope}{a boolean to manage first or second slope in the roughness against lenghtscale curve}
}
\value{
a \code{\link{AFMImagePSDSlopesAnalysis}}
}
\description{
\code{\link{getIntersectionForRoughnessAgainstLengthscale}} get the intersection between tangente and plateau
}
\author{
M.Beauvais
}
