% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getIntersectionPointWithBorder}
\alias{getIntersectionPointWithBorder}
\title{getIntersectionPointWithBorder to be described}
\usage{
getIntersectionPointWithBorder(AFMImage, center, r, deg)
}
\arguments{
\item{AFMImage}{a \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{center}{center}

\item{r}{radius}

\item{deg}{degree}
}
\description{
\code{getIntersectionPointWithBorder} return a data.table
}
\author{
M.Beauvais
}
