% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getListOfDiameters}
\alias{getListOfDiameters}
\title{getListOfDiameters}
\usage{
getListOfDiameters(g)
}
\arguments{
\item{g}{list of igraph networks}
}
\description{
\code{getListOfDiameters} return
}
\author{
M.Beauvais
}
