% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getMaxCircleMatrix}
\alias{getMaxCircleMatrix}
\title{getMaxCircleMatrix}
\usage{
getMaxCircleMatrix(..., newCircleAFMImage, CIRCLE_RADIUS_INIT)
}
\arguments{
\item{...}{cl: a cluster object from the parallel package}

\item{newCircleAFMImage}{a \code{\link{AFMImage}}}

\item{CIRCLE_RADIUS_INIT}{CIRCLE_RADIUS_INIT}
}
\value{
res a matrix
}
\description{
for each pixel of the image,
 if the pixel is not empty
try to place one circle
start with biggets circle
as soon as a circle is found the circle, the pixel is associated with with the circle raidus
}
\author{
M.Beauvais
}
