% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getNetworkGridLayout}
\alias{getNetworkGridLayout}
\title{Get getNetworkGridLayout}
\usage{
getNetworkGridLayout(AFMImage, vId)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{vId}{the vertex id}
}
\description{
\code{getNetworkGridLayout} return a list x,y coordinates
}
\author{
M.Beauvais
}

