% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getSurroundingVertexesList}
\alias{getSurroundingVertexesList}
\title{Get surrounding vertices from x,y coordinates}
\usage{
getSurroundingVertexesList(AFMImage, x, y)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{x}{coordinates in x axis}

\item{y}{coordinates in y axis}
}
\description{
\code{getSurroundingVertexesList} return the vertexId
}
\author{
M.Beauvais
}
