% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getTopologyAFMImage}
\alias{getTopologyAFMImage}
\title{Calculate topology image (TBC)}
\usage{
getTopologyAFMImage(BinaryAFMImage, AFMImageNetworksAnalysis)
}
\arguments{
\item{BinaryAFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy in a binary format 0 or 1 values for heigths}

\item{AFMImageNetworksAnalysis}{an \code{\link{AFMImageNetworksAnalysis}} from Atomic Force Microscopy}
}
\description{
\code{getTopologyAFMImage} return the global topological distance
}
\author{
M.Beauvais
}
