% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getTriangle}
\alias{getTriangle}
\title{get a triangle starting from center, two segments of length r with angles deg1 and deg2}
\usage{
getTriangle(AFMImage, center, r, deg1, deg2)
}
\arguments{
\item{AFMImage}{a \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{center}{center}

\item{r}{length of segment}

\item{deg1}{angle 1}

\item{deg2}{angel 2}
}
\description{
\code{getTriangle} return a data.table points of a triangle
}
\author{
M.Beauvais
}
