% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getVertexId}
\alias{getVertexId}
\title{Get vertex id from x,y coordinates}
\usage{
getVertexId(AFMImage, x, y)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{x}{coordinates in x axis}

\item{y}{coordinates in y axis}
}
\description{
\code{getVertexId} return the vertexId
}
\author{
M.Beauvais
}
