% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{gridIgraphPlot}
\alias{gridIgraphPlot}
\title{gridIgraphPlot}
\usage{
gridIgraphPlot(AFMImage, g)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{g}{the networks}
}
\description{
\code{gridIgraphPlot} return TRUE if vertex is adjacent to a better vertex
}
\author{
M.Beauvais
}
