% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{identifyEdgesFromCircles}
\alias{identifyEdgesFromCircles}
\title{display the network of nodes and edges}
\usage{
identifyEdgesFromCircles(AFMImageNetworksAnalysis, MAX_DISTANCE = 40,
  ANGLE_MIN = (pi/3.25))
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{MAX_DISTANCE}{the maximum distance between nodes to check if nodes are connected. Default value is 40.}

\item{ANGLE_MIN}{the minimum angle between edges to accept or reject an edge}
}
\description{
display the network of nodes and edges
}
\author{
M.Beauvais
}

