% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{identifyMaxCircleRadius}
\alias{identifyMaxCircleRadius}
\title{identifyMaxCircleRadius}
\usage{
identifyMaxCircleRadius(
  i,
  allXY,
  newCircleAFMImage,
  binaryAFMImageMatrix,
  maxCircleRadiusMatrix,
  circleRadius,
  circlenm
)
}
\arguments{
\item{i}{an integer}

\item{allXY}{combinations of ?}

\item{newCircleAFMImage}{a \code{\link{AFMImage}}}

\item{binaryAFMImageMatrix}{a \code{\link{AFMImage}}}

\item{maxCircleRadiusMatrix}{a matrix}

\item{circleRadius}{a vector of radius ?}

\item{circlenm}{a ?}
}
\value{
a data table with x,y,radius columns
}
\description{
identifyMaxCircleRadius
}
\author{
M.Beauvais
}
