% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{identifyNodesAndEdges}
\alias{identifyNodesAndEdges}
\title{identifyNodesAndEdges}
\usage{
identifyNodesAndEdges(..., AFMImageNetworksAnalysis, maxHeight)
}
\arguments{
\item{...}{cl: a cluster object from the parallel package}

\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{maxHeight}{a double for filtering the heights - upper to this height the heights are set to zero}
}
\value{
AFMImageNetworksAnalysis a \code{\link{AFMImageNetworksAnalysis}}
}
\description{
find nodes and edges
}
\author{
M.Beauvais
}
