% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{identifyNodesWithCircles}
\alias{identifyNodesWithCircles}
\title{identify largest circles in binary image}
\usage{
identifyNodesWithCircles(..., AFMImageNetworksAnalysis)
}
\arguments{
\item{...}{cl: a cluster object from the parallel package}

\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}
}
\value{
AFMImageNetworksAnalysis the \code{\link{AFMImageNetworksAnalysis}} instance
}
\description{
\code{identifyNodesWithCircles} return TRUE if vertex is adjacent to a better vertex
}
\author{
M.Beauvais
}
