% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{isAdjacentToBetterVertex}
\alias{isAdjacentToBetterVertex}
\title{isAdjacentToBetterVertex}
\usage{
isAdjacentToBetterVertex(AFMImage, x, y)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{x}{coordinates in x axis}

\item{y}{coordinates in y axis}
}
\description{
\code{isAdjacentToBetterVertex} return TRUE if vertex is adjacent to a better vertex
}
\author{
M.Beauvais
}
