% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{isAngleBetweenEdgesAlwaysSuperiorToMinAngle}
\alias{isAngleBetweenEdgesAlwaysSuperiorToMinAngle}
\title{check if all the angles between one edge and a list of edges is superior to a specified value.}
\usage{
isAngleBetweenEdgesAlwaysSuperiorToMinAngle(
  binaryAFMImage,
  edge1,
  edges2,
  minAngle
)
}
\arguments{
\item{binaryAFMImage}{a binary \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{edge1}{one edge}

\item{edges2}{list of edges}

\item{minAngle}{the minimum angle value}
}
\value{
TRUE if all the angle are superior to the specified value
}
\description{
check if all the angles between one edge and a list of edges is superior to a specified value.
}
\author{
M.Beauvais
}
