% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMGaussianMixAnalyser.R
\name{loglike.normalmix}
\alias{loglike.normalmix}
\title{loglike sum of density of a mixture of normals}
\usage{
loglike.normalmix(x, mixture)
}
\arguments{
\item{x}{a vector of quantiles}

\item{mixture}{a gaussian mixture}
}
\description{
loglike sum of density of a mixture of normals
}
