% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMVariogramAnalyser.R
\name{omniVariogramSlopeAnalysis-class}
\alias{omniVariogramSlopeAnalysis-class}
\alias{omniVariogramSlopeAnalysis}
\alias{initialize,omniVariogramSlopeAnalysis-method}
\title{AFM Image log-log experimental variogram slope analysis}
\usage{
omniVariogramSlopeAnalysis()

\S4method{initialize}{omniVariogramSlopeAnalysis}(.Object)

omniVariogramSlopeAnalysis()
}
\arguments{
\item{.Object}{an omniVariogramSlopeAnalysis object}
}
\description{
\code{omniVariogramSlopeAnalysis} stores the analysis of the second slope in roughness against lenghtscale
}
\section{Slots}{

\describe{
\item{\code{intersection_sill}}{to be removed ?}

\item{\code{sill}}{to be removed ?}

\item{\code{slope}}{to be removed ?}

\item{\code{yintersept}}{to be removed ?}
}}

\author{
M.Beauvais
}
