% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AFMPSDAnalyser.R
\name{performAllPSDCalculation}
\alias{performAllPSDCalculation}
\title{Perform all the calculation for PSD exploitation}
\usage{
performAllPSDCalculation(AFMImagePSDAnalysis, AFMImage)
}
\arguments{
\item{AFMImagePSDAnalysis}{an \code{\link{AFMImagePSDAnalysis}} to manage and store the results of PSD analysis}

\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}
}
\description{
\code{\link{performAllPSDCalculation}} perform all the calculation for PSD exploitation
}
\examples{
library(AFM)

data(AFMImageOfNormallyDistributedHeights)
}
\author{
M.Beauvais
}

