% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMGaussianMixAnalyser.R
\name{pnormmix}
\alias{pnormmix}
\title{pnormmix distribution of a mixture of normals}
\usage{
pnormmix(q, mixture)
}
\arguments{
\item{q}{a vector of quantiles}

\item{mixture}{a gaussian mixture}
}
\description{
pnormmix distribution of a mixture of normals
}
