% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMReportMaker.R
\name{printVariogramModelEvaluations}
\alias{printVariogramModelEvaluations}
\title{printVariogramModelEvaluations}
\usage{
printVariogramModelEvaluations(
  AFMImageAnalyser,
  sampleDT,
  numberOfModelsPerPage
)
}
\arguments{
\item{AFMImageAnalyser}{an \code{\link{AFMImageAnalyser}} to be used to produce report}

\item{sampleDT}{a data.table containg the evaluation information}

\item{numberOfModelsPerPage}{numeric to specify the number of model evaluations per pages}
}
\description{
\code{printVariogramModelEvaluations} generates a graphic element containing the evaluation of all variogram models
}
\author{
M.Beauvais
}
