% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMImageAnalyser.R
\name{putAnalysisOnDisk}
\alias{putAnalysisOnDisk}
\alias{putAnalysisOnDisk,AFMImageAnalyser-method}
\title{Export all data from an analysis of an AFM Image as rdata file}
\usage{
putAnalysisOnDisk(AFMImageAnalyser, AFMImage)

\S4method{putAnalysisOnDisk}{AFMImageAnalyser}(AFMImageAnalyser, AFMImage)
}
\arguments{
\item{AFMImageAnalyser}{an \code{\link{AFMImageAnalyser}}}

\item{AFMImage}{an \code{\link{AFMImage}}}
}
\description{
A function to export to all the data from all analysis of an \code{\link{AFMImage}} and put them on disk as rdata file
}
\author{
M.Beauvais
}
