% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{removeLonguestEdge}
\alias{removeLonguestEdge}
\title{removeLonguestEdge}
\usage{
removeLonguestEdge(k, res, sides, myRes, vertex1)
}
\arguments{
\item{k}{an integer}

\item{res}{res ?}

\item{sides}{data.table}

\item{myRes}{data.table?}

\item{vertex1}{a vertex ?}
}
\value{
a data.table with from, to
}
\description{
Find and remove the longuest edge if it is unique
}
\author{
M.Beauvais
}
