% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{removeNode}
\alias{removeNode}
\title{removeNode}
\usage{
removeNode(circleAFMImage, nodeDT)
}
\arguments{
\item{circleAFMImage}{a \code{\link{AFMImage}}}

\item{nodeDT}{a data.table lon lat circleRadius}
}
\value{
an \code{\link{AFMImage}}
}
\description{
remove a node from an AFMImage
}
\author{
M.Beauvais
}
