% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAFMApp.R
\name{runAFMApp}
\alias{runAFMApp}
\title{Launch the AFM shiny application}
\usage{
runAFMApp()
}
\description{
Launch the AFM shiny graphical interface to access most of the fonctionalities of the AFM library
}
\examples{
\dontrun{
install.packages("AFM")
AFM::runAFMApp()
}
}
\author{
M.Beauvais
}
