% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMPSDAnalyser.R
\name{saveOnDiskIntersectionForRoughnessAgainstLengthscale}
\alias{saveOnDiskIntersectionForRoughnessAgainstLengthscale}
\title{save an image of the roughness against lenghtscale calculations}
\usage{
saveOnDiskIntersectionForRoughnessAgainstLengthscale(
  AFMImageAnalyser,
  exportDirectory
)
}
\arguments{
\item{AFMImageAnalyser}{an \code{\link{AFMImageAnalyser}} to get Roughness against lenghtscale calculation}

\item{exportDirectory}{a directory on the file system}
}
\description{
\code{\link{saveOnDiskIntersectionForRoughnessAgainstLengthscale}} save an image of the roughness against lenghtscale calculations
}
\author{
M.Beauvais
}
