% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{simplifyNetwork}
\alias{simplifyNetwork}
\title{simplifyNetwork}
\usage{
simplifyNetwork(..., allVertices, allEdges)
}
\arguments{
\item{...}{cl: a cluster object from the parallel package}

\item{allVertices}{a data.table of vertices}

\item{allEdges}{a data.table of edges}
}
\value{
a data.table of edges
}
\description{
simplify the network keeping only the important edges
}
\author{
M.Beauvais
}
