% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{thinImage}
\alias{thinImage}
\title{thin an Image in matrix format}
\usage{
thinImage(imageMatrix)
}
\arguments{
\item{imageMatrix}{a matrix of an AFM image}
}
\description{
thin an Image in matrix format
}
\author{
M.Beauvais
}
