% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AFMImageAnalyser.R
\docType{methods}
\name{initialize,AFMImageAnalyser-method}
\alias{initialize,AFMImageAnalyser-method}
\title{Constructor method of AFMImageAnalyser Class.}
\usage{
\S4method{initialize}{AFMImageAnalyser}(.Object, variogramAnalysis, psdAnalysis,
  fdAnalysis, mean, variance, TotalRrms, Ra, fullfilename)
}
\arguments{
\item{.Object}{an AFMImageAnalyser object}

\item{variogramAnalysis}{\code{\link{AFMImageVariogramAnalysis}}}

\item{psdAnalysis}{\code{\link{AFMImagePSDAnalysis}}}

\item{fdAnalysis}{\code{\link{AFMImageFractalDimensionsAnalysis}}}

\item{mean}{the mean of heights of the \code{\link{AFMImage}}}

\item{variance}{the variance of heights of the \code{\link{AFMImage}}}

\item{TotalRrms}{the total Root Mean Square Roughness of the \code{\link{AFMImage}} calculated from variance}

\item{Ra}{mean roughness or mean of absolute values of heights}

\item{fullfilename}{to be removed?}
}
\description{
Constructor method of AFMImageAnalyser Class.
}

