% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AFMFractalDimensionAnalyser.R
\docType{methods}
\name{AFMImageFractalDimensionMethod-class}
\alias{AFMImageFractalDimensionMethod}
\alias{AFMImageFractalDimensionMethod-class}
\alias{initialize,AFMImageFractalDimensionMethod-method}
\title{AFM image fractal dimension method class}
\usage{
AFMImageFractalDimensionMethod(fd_method, fd, fd_scale)

\S4method{initialize}{AFMImageFractalDimensionMethod}(.Object, fd_method, fd,
  fd_scale)

AFMImageFractalDimensionMethod(fd_method, fd, fd_scale)
}
\arguments{
\item{fd_method}{Two dimensional function names used to evaluate the fractal dimension and fractal scale}

\item{fd}{the value of the fractal dimension}

\item{fd_scale}{the value of the fractal scale}

\item{.Object}{an AFMImageFractalDimensionMethod object}
}
\description{
\code{AFMImageFractalDimensionMethod} stores calculation from one fractal dimension method

Constructor method of AFMImageFractalDimensionMethod Class.

Wrapper function AFMImageFractalDimensionMethod
}
\section{Slots}{

\describe{
\item{\code{fd_method}}{Two dimensional function names used to evaluate the fractal dimension and fractal scale}

\item{\code{fd}}{the value of the fractal dimension}

\item{\code{fd_scale}}{the value of the fractal scale}
}}
\author{
M.Beauvais
}
\seealso{
\code{\link{fractaldim}}
}

