% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\docType{methods}
\name{AFMImageNetworksAnalysis-class}
\alias{AFMImageNetworksAnalysis}
\alias{AFMImageNetworksAnalysis-class}
\alias{initialize,AFMImageNetworksAnalysis-method}
\title{AFM image networks analysis class}
\usage{
AFMImageNetworksAnalysis()

\S4method{initialize}{AFMImageNetworksAnalysis}(.Object, originalGraph,
  skeletonGraph, heightNetworksslider, filterNetworkssliderMin,
  filterNetworkssliderMax, libVersion)

AFMImageNetworksAnalysis()
}
\arguments{
\item{.Object}{an AFMImageNetworksAnalysis Class}

\item{originalGraph}{a list of \code{\link{igraph}}}

\item{skeletonGraph}{a list of \code{\link{igraph}}}

\item{heightNetworksslider}{used multiplier of heights to facilitate analysis}

\item{filterNetworkssliderMin}{used filter minimum value to facilitate analysis}

\item{filterNetworkssliderMax}{used filter maximum value to facilitate analysis}

\item{libVersion}{version of the AFM library used to perform the analysis}
}
\description{
A S4 class to handle the networks calculation

Constructor method of AFMImageNetworksAnalysis Class.

Wrapper function AFMImageNetworksAnalysis
}
\section{Slots}{

\describe{
\item{\code{originalGraph}}{a list of \code{\link{igraph}}}

\item{\code{skeletonGraph}}{a list of \code{\link{igraph}}}

\item{\code{heightNetworksslider}}{used multiplier of heights to facilitate analysis}

\item{\code{filterNetworkssliderMin}}{used filter minimum value to facilitate analysis}

\item{\code{filterNetworkssliderMax}}{used filter maximum value to facilitate analysis}

\item{\code{libVersion}}{version of the AFM library used to perform the analysis}

\item{\code{updateProgress}}{a function to update a graphical user interface}
}}
\author{
M.Beauvais
}

