% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AFMPSDAnalyser.R
\docType{methods}
\name{AFMImagePSDAnalysis-class}
\alias{AFMImagePSDAnalysis}
\alias{AFMImagePSDAnalysis-class}
\alias{initialize,AFMImagePSDAnalysis-method}
\alias{intersections}
\alias{intersections,AFMImagePSDAnalysis-method}
\alias{roughnessAgainstLengthscale}
\alias{roughnessAgainstLengthscale,AFMImagePSDAnalysis-method}
\title{AFM image Power Spectrum Density analysis class}
\usage{
AFMImagePSDAnalysis()

\S4method{initialize}{AFMImagePSDAnalysis}(.Object)

AFMImagePSDAnalysis()

roughnessAgainstLengthscale(object)

\S4method{roughnessAgainstLengthscale}{AFMImagePSDAnalysis}(object)

intersections(object)

\S4method{intersections}{AFMImagePSDAnalysis}(object)
}
\arguments{
\item{.Object}{an AFMImagePSDAnalysis object}

\item{object}{a \code{\link{AFMImagePSDAnalysis}}}
}
\description{
\code{AFMImagePSDAnalysis} handles an \code{\link{AFMImage}} roughness against lenghscale analysis

Constructor method of AFMImagePSDAnalysis Class.

Wrapper function AFMImagePSDAnalysis

Method \code{roughnessAgainstLengthscale} returns a data.table of roughness vs. Lengthscale

Method \code{intersections} returns a intersection numeric value
}
\section{Slots}{

\describe{
\item{\code{roughnessAgainstLengthscale}}{a data.table to store the roughness against lengthscale data}

\item{\code{intersections}}{a list to store the lengthscales values as the intersections between slopes and the sill in roughness against lenghscale graph}
}}
\author{
M.Beauvais
}

