% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMPSDAnalyser.R
\docType{methods}
\name{AFMImagePSDSlopesAnalysis-class}
\alias{AFMImagePSDSlopesAnalysis}
\alias{AFMImagePSDSlopesAnalysis-class}
\alias{initialize,AFMImagePSDSlopesAnalysis-method}
\title{AFM Image psd slope analysis}
\usage{
AFMImagePSDSlopesAnalysis()

\S4method{initialize}{AFMImagePSDSlopesAnalysis}(.Object)

AFMImagePSDSlopesAnalysis()
}
\arguments{
\item{.Object}{an AFMImagePSDSlopesAnalysis object}
}
\description{
\code{AFMImagePSDSlopesAnalysis} stores the analysis of the second slope in roughness against lenghtscale

Constructor method of AFMImagePSDSlopesAnalysis Class.

Wrapper function AFMImagePSDSlopesAnalysis
}
\section{Slots}{

\describe{
\item{\code{lc}}{to be removed ?}

\item{\code{wsat}}{to be removed ?}

\item{\code{slope}}{to be removed ?}

\item{\code{yintersept}}{to be removed ?}
}}
\author{
M.Beauvais
}

