% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculatePhysicalDistanceFromPath}
\alias{calculatePhysicalDistanceFromPath}
\title{calculate the physical distances between nodes}
\usage{
calculatePhysicalDistanceFromPath(pathVidVector, hscale, vscale)
}
\arguments{
\item{pathVidVector}{a network path}

\item{hscale}{the hscale of the \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{vscale}{the vscale of the \code{\link{AFMImage}} from Atomic Force Microscopy}
}
\value{
the physical distance the extrmities of the path
}
\description{
calculate the physical distances between nodes
}
\author{
M.Beauvais
}
