% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{calculateShortestPaths}
\alias{calculateShortestPaths}
\title{calculate the  shortest path between adjacent nodes
Calculateshrotest path between all nodes of degree superior to 3
identify nodes that are connected by nodes which degrees is inferior to 2
and calculate the distance between these nodes}
\usage{
calculateShortestPaths(AFMImageNetworksAnalysis, fusioned = TRUE)
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{fusioned}{TRUE if graph is created from fusioned nodes}
}
\description{
calculate the  shortest path between adjacent nodes
Calculateshrotest path between all nodes of degree superior to 3
identify nodes that are connected by nodes which degrees is inferior to 2
and calculate the distance between these nodes
}
\author{
M.Beauvais
}

