% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{createGraph}
\alias{createGraph}
\title{create the igraph graph from the fusionned nodes and the edges}
\usage{
createGraph(AFMImageNetworksAnalysis, fusioned = TRUE)
}
\arguments{
\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{fusioned}{TRUE if graph is created from fusioned nodes}
}
\description{
create the igraph graph from the fusionned nodes and the edges
}
\author{
M.Beauvais
}

