% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{fusionCloseNodes}
\alias{fusionCloseNodes}
\title{fusion the nodes that are intersecting}
\usage{
fusionCloseNodes(AFMImageNetworksAnalysis)
}
\arguments{
\item{AFMImageNetworksAnalysis}{the AFMImageNetworksAnalysis instance}
}
\value{
a list of edges with fusioned nodes
}
\description{
fusion the nodes that are intersecting
}
\author{
M.Beauvais
}

