% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getCoordinatesFromVertexId}
\alias{getCoordinatesFromVertexId}
\title{Get x,y coordinates from vertex id}
\usage{
getCoordinatesFromVertexId(vId)
}
\arguments{
\item{vId}{the vertex id}
}
\description{
\code{getCoordinatesFromVertexId} return a list x,y coordinates
}
\author{
M.Beauvais
}
