% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{getNetworkGridLayout}
\alias{getNetworkGridLayout}
\title{#' @export
getCoordinatesFromVertexId2<-function(AFMImage, vId) {
  vertexId<-as.numeric(vId)
  y<-floor(vertexId/HASHSIZE)
  x<-vertexId-y*HASHSIZE
  return(data.table(vId=vId, coords.x1=x,coords.x2=y))
}
Get getNetworkGridLayout}
\usage{
getNetworkGridLayout(AFMImage, vId)
}
\arguments{
\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}

\item{vId}{the vertex id}
}
\description{
\code{getNetworkGridLayout} return a list x,y coordinates
}
\author{
M.Beauvais
}
