% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{identifyEdgesFromCircles}
\alias{identifyEdgesFromCircles}
\title{display the network of nodes and edges}
\usage{
identifyEdgesFromCircles(..., AFMImageNetworksAnalysis, MAX_DISTANCE = 40)
}
\arguments{
\item{...}{cl: a cluster object from the parallel package}

\item{AFMImageNetworksAnalysis}{a \code{\link{AFMImageNetworksAnalysis}}}

\item{MAX_DISTANCE}{the maximum distance between nodes to check if nodes are connected. Default value is 40.}
}
\description{
display the network of nodes and edges
}
\author{
M.Beauvais
}
