% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMNetworksAnalyser.R
\name{transformAFMImageForNetworkAnalysis}
\alias{transformAFMImageForNetworkAnalysis}
\alias{transformAFMImageForNetworkAnalysis,AFMImageNetworksAnalysis-method}
\alias{transformAFMImageForNetworkAnalysis,AFMImage-method}
\title{Multiply, filter the heights and make a binary AFMImage from the transformed AFMImage}
\usage{
transformAFMImageForNetworkAnalysis(AFMImageNetworksAnalysis, AFMImage)

\S4method{transformAFMImageForNetworkAnalysis}{AFMImageNetworksAnalysis}(AFMImageNetworksAnalysis, AFMImage)
}
\arguments{
\item{AFMImageNetworksAnalysis}{n \code{\link{AFMImageNetworksAnalysis}} to store the results of the transformation}

\item{AFMImage}{an \code{\link{AFMImage}} from Atomic Force Microscopy}
}
\description{
\code{transformAFMImageForNetworkAnalysis} update  \code{\link{AFMImageNetworksAnalysis}} making a binary AFMImage
}
\author{
M.Beauvais
}
